broke out properties to their own files. Added documentation
